backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
plotly>=5.3.0
python-dotenv>=0.19.0

# Test dependencies
pytest>=7.0.0
httpx>=0.24.0  # For async HTTP testing
pytest-asyncio>=0.21.0  # For async test support 